added workaround for bug in Node v8.9.4 and below
note that this assumes the issue will be fixed in the upcoming Node
release
