+++ lib/manager.js
@@ -1,6 +1,7 @@
 "use strict";
 
 let Manifest = require("./manifest");
+let resolveModulePath = require("./util/resolve");
 let createFile = require("./util/files");
 let { generateFingerprint, repr } = require("./util");
 let path = require("path");
@@ -84,23 +85,3 @@ module.exports = class AssetManager {
 		}
 	}
 };
-
-function resolveModulePath(filepath, rootDir) {
-	// older versions of Node do not support `require.resolve`'s `paths` option
-	let legacy = !require.resolve.paths;
-	if(legacy) {
-		legacy = process.env.NODE_PATH; // cache previous value
-		rootDir = rootDir.replace(/\/{1,}$/, ""); // strip trailing slashes, to be safe
-		process.env.NODE_PATH = `${rootDir}:${rootDir}/node_modules`;
-		require("module").Module._initPaths();
-	}
-
-	let res = require.resolve(filepath, { paths: [rootDir] });
-
-	if(legacy) { // restore previous environment
-		process.env.NODE_PATH = legacy;
-		require("module").Module._initPaths();
-	}
-
-	return res;
-}
